Model split into components
