--- .idea/binnacle-dart.iml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/.pub" />
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/build" />
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/ios/.symlinks/plugins/flutter_compass/.pub" />
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/ios/.symlinks/plugins/flutter_compass/build" />
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/ios/.symlinks/plugins/flutter_compass/example/.pub" />
-      <excludeFolder url="file://$MODULE_DIR$/binnacle/ios/.symlinks/plugins/flutter_compass/example/build" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
